Let realloc handle NULL.  Check for integer overflow and realloc failure.  Avoid passing negative integers to realloc by changing variables to size_t.
Initial diff from Charles Longeau.

OK millert.

+++ usr.bin/systat/extern.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: extern.h,v 1.17 2007/02/24 09:45:27 otto Exp $	*/
+/*	$OpenBSD: extern.h,v 1.18 2007/04/24 06:32:08 tedu Exp $	*/
 /*	$NetBSD: extern.h,v 1.3 1996/05/10 23:16:34 thorpej Exp $	*/
 
 /*-
@@ -49,8 +49,8 @@ extern int	CMDLINE;
 extern int	dk_ndrive;
 extern int	hz, stathz;
 extern double	naptime;
-extern int	nhosts;
-extern int	nports;
+extern size_t	nhosts;
+extern size_t	nports;
 extern int	protos;
 extern int	verbose;
 extern int	nflag;
